% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_world}
\alias{hello_world}
\title{Return a friendly message}
\usage{
hello_world(world = "World")
}
\arguments{
\item{world}{A character vector to whom the message is directed}
}
\value{
A character vector
}
\description{
Return a friendly message
}
